requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
scikit-learn
pandas
opencv-python-headless